reverted changes in onlinebeamspotproducer
