+++ RecoVertex/BeamSpotProducer/plugins/BeamSpotOnlineProducer.cc
@@ -1,46 +1,14 @@
-/**_________________________________________________________________
-   class:   BeamSpotOnlineProducer.h
-   package: RecoVertex/BeamSpotProducer
 
-
-
- author: Francisco Yumiceva, Fermilab (yumiceva@fnal.gov)
-
-
-________________________________________________________________**/
-
+#include "RecoVertex/BeamSpotProducer/interface/BeamSpotOnlineProducer.h"
 #include "DataFormats/BeamSpot/interface/BeamSpot.h"
-#include "DataFormats/Common/interface/Handle.h"
-#include "DataFormats/L1GlobalTrigger/interface/L1GlobalTriggerEvmReadoutRecord.h"
-#include "DataFormats/Scalers/interface/BeamSpotOnline.h"
-#include "CondFormats/BeamSpotObjects/interface/BeamSpotObjects.h"
-#include "CondFormats/DataRecord/interface/BeamSpotObjectsRcd.h"
+
 #include "FWCore/Framework/interface/ESHandle.h"
-#include "FWCore/Framework/interface/Event.h"
 #include "FWCore/Framework/interface/EventSetup.h"
-#include "FWCore/Framework/interface/stream/EDProducer.h"
-#include "FWCore/MessageLogger/interface/MessageLogger.h"
-#include "FWCore/ParameterSet/interface/ParameterSet.h"
-#include "FWCore/Utilities/interface/ESGetToken.h"
+#include "FWCore/Framework/interface/IOVSyncValue.h"
+#include "CondFormats/DataRecord/interface/BeamSpotObjectsRcd.h"
+#include "CondFormats/BeamSpotObjects/interface/BeamSpotObjects.h"
 
-class BeamSpotOnlineProducer : public edm::stream::EDProducer<> {
-public:
-  /// constructor
-  explicit BeamSpotOnlineProducer(const edm::ParameterSet& iConf);
-
-  /// produce a beam spot class
-  void produce(edm::Event& iEvent, const edm::EventSetup& iSetup) override;
-
-private:
-  const bool changeFrame_;
-  const double theMaxZ, theSetSigmaZ;
-  double theMaxR2;
-  const edm::EDGetTokenT<BeamSpotOnlineCollection> scalerToken_;
-  const edm::EDGetTokenT<L1GlobalTriggerEvmReadoutRecord> l1GtEvmReadoutRecordToken_;
-  const edm::ESGetToken<BeamSpotObjects, BeamSpotObjectsRcd> beamToken_;
-
-  const unsigned int theBeamShoutMode;
-};
+#include "FWCore/Framework/interface/MakerMacros.h"
 
 using namespace edm;
 
@@ -50,13 +18,14 @@ BeamSpotOnlineProducer::BeamSpotOnlineProducer(const ParameterSet& iconf)
       theSetSigmaZ(iconf.getParameter<double>("setSigmaZ")),
       scalerToken_(consumes<BeamSpotOnlineCollection>(iconf.getParameter<InputTag>("src"))),
       l1GtEvmReadoutRecordToken_(consumes<L1GlobalTriggerEvmReadoutRecord>(iconf.getParameter<InputTag>("gtEvmLabel"))),
-      beamToken_(esConsumes<BeamSpotObjects, BeamSpotObjectsRcd>()),
       theBeamShoutMode(iconf.getUntrackedParameter<unsigned int>("beamMode", 11)) {
   theMaxR2 = iconf.getParameter<double>("maxRadius");
   theMaxR2 *= theMaxR2;
 
   produces<reco::BeamSpot>();
 }
+
+BeamSpotOnlineProducer::~BeamSpotOnlineProducer() {}
 
 void BeamSpotOnlineProducer::produce(Event& iEvent, const EventSetup& iSetup) {
   //shout MODE only in stable beam
@@ -137,7 +106,8 @@ void BeamSpotOnlineProducer::produce(Event& iEvent, const EventSetup& iSetup) {
   }
 
   if (fallBackToDB) {
-    edm::ESHandle<BeamSpotObjects> beamhandle = iSetup.getHandle(beamToken_);
+    edm::ESHandle<BeamSpotObjects> beamhandle;
+    iSetup.get<BeamSpotObjectsRcd>().get(beamhandle);
     const BeamSpotObjects* spotDB = beamhandle.product();
 
     // translate from BeamSpotObjects to reco::BeamSpot
@@ -165,5 +135,4 @@ void BeamSpotOnlineProducer::produce(Event& iEvent, const EventSetup& iSetup) {
   iEvent.put(std::move(result));
 }
 
-#include "FWCore/Framework/interface/MakerMacros.h"
 DEFINE_FWK_MODULE(BeamSpotOnlineProducer);
